Mergerd project Popup and projectCards to avoid duplicate properties
