+++ assets/css/styles.css
@@ -738,12 +738,12 @@ HAMBURGER MENU DISPLACED DUE TO LINTERS ERROR
 
 .popup-modal {
   position: absolute;
-  top: 0;
+  top: 50%;
   left: 0;
   width: 100%;
   height: 100%;
   background: var(--image-placeholder-color);
-  z-index: 8;
+  z-index: 100;
 }
 
 .project-popup {
@@ -762,7 +762,7 @@ HAMBURGER MENU DISPLACED DUE TO LINTERS ERROR
 
 .project-popup-desktop-img {
   width: 100%;
-  height: 80%;
+  height: 60%;
   margin-top: 3%;
   border-radius: 1%;
   border: 1px solid white;
@@ -772,11 +772,11 @@ HAMBURGER MENU DISPLACED DUE TO LINTERS ERROR
 
 .project-popup-mobile-img {
   width: 100% !important;
-  height: 60%;
+  height: 40%;
   margin-top: 3%;
   border-radius: 1%;
   border: 1px solid white;
-  /*background: url(../../images/popup/SnapshootPortfolioMobile.svg) no-repeat; /* TO REMOVE AFTER FIXING */
+  /* background: url(../../images/popup/SnapshootPortfolioMobile.svg) no-repeat; /*TO REMOVE AFTER FIXING */
   background-size: 100%; /* TO REMOVE AFTER FIXING */
 }
 
@@ -796,6 +796,10 @@ HAMBURGER MENU DISPLACED DUE TO LINTERS ERROR
   visibility: hidden;
   z-index: 1000;
   background-color: var(--image-placeholder-color);
+}
+
+.open-project-popup .active {
+  visibility: visible;
 }
 
 .project-card-js {
