+++ index.html
@@ -285,7 +285,7 @@
   </footer>
 
   <script type="module" src="assets/js/hamburger.js"></script>
-  <script type="module" src="assets/js/projectCards.js"></script>
+  <script type="module" src="assets/js/popup.js"></script>
 
 </body>
 </html>
